Move Generator Screen widgets to separate files
