+++ lib/ui/screens/generator_screen.dart
@@ -5,7 +5,8 @@ import 'package:flutter_gradient_generator/data/app_strings.dart';
 import 'package:flutter_gradient_generator/enums/gradient_direction.dart';
 import 'package:flutter_gradient_generator/enums/gradient_style.dart';
 import 'package:flutter_gradient_generator/models/abstract_gradient.dart';
-import 'package:flutter_vector_icons/flutter_vector_icons.dart';
+import 'package:flutter_gradient_generator/ui/widgets/direction_widget.dart';
+import 'package:flutter_gradient_generator/ui/widgets/get_gradient_button.dart';
 
 class GeneratorScreen extends StatelessWidget {
   final AbstractGradient gradient;
@@ -168,201 +169,4 @@ class GeneratorScreen extends StatelessWidget {
   }
 }
 
-class DirectionWidget extends StatelessWidget {
-  final GradientStyle gradientStyle;
-  final GradientDirection selectedGradientDirection;
-  final void Function(GradientDirection) onGradientDirectionChanged;
 
-  DirectionWidget(
-      {Key? key,
-      required this.gradientStyle,
-      required this.selectedGradientDirection,
-      required this.onGradientDirectionChanged})
-      : super(key: key);
-
-  final int circleDirectionIconSetNumber = 1;
-  final int circleDirectionIconNumberInSet = 1;
-
-  final List<Map<GradientDirection, IconData>> iconSetList = [
-    {
-      GradientDirection.topLeft: MaterialCommunityIcons.arrow_top_left,
-      GradientDirection.topCenter: MaterialCommunityIcons.arrow_up,
-      GradientDirection.topRight: MaterialCommunityIcons.arrow_top_right,
-    },
-    {
-      GradientDirection.centerLeft: MaterialCommunityIcons.arrow_left,
-      GradientDirection.center: MaterialCommunityIcons.circle_outline,
-      GradientDirection.centerRight: MaterialCommunityIcons.arrow_right,
-    },
-    {
-      GradientDirection.bottomLeft: MaterialCommunityIcons.arrow_bottom_left,
-      GradientDirection.bottomCenter: MaterialCommunityIcons.arrow_down,
-      GradientDirection.bottomRight: MaterialCommunityIcons.arrow_bottom_right,
-    }
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: iconSetList.map(
-            (Map<GradientDirection, IconData> gradientDirectionToIconSetMap) {
-          final int iconSetIndex =
-              iconSetList.indexOf(gradientDirectionToIconSetMap);
-          final int firstIconSetIndex = 0;
-
-          return Column(
-            children: [
-              if (iconSetIndex != firstIconSetIndex) SizedBox(height: 8.0),
-              Row(
-                  children: gradientDirectionToIconSetMap.values.map((icon) {
-                final int iconIndex =
-                    gradientDirectionToIconSetMap.values.toList().indexOf(icon);
-                final GradientDirection gradientDirection =
-                    gradientDirectionToIconSetMap.keys.elementAt(iconIndex);
-
-                final int firstIconIndex = 0;
-
-                final bool isCircleRadialButton =
-                    iconSetIndex == circleDirectionIconSetNumber &&
-                        iconIndex == circleDirectionIconNumberInSet;
-
-                return Expanded(
-                  child: Row(
-                    children: [
-                      if (iconIndex != firstIconIndex) SizedBox(width: 8.0),
-                      Expanded(
-                        child: Visibility(
-                          child: DirectionButton(
-                              icon: icon,
-                              gradientDirection: gradientDirection,
-                              isSelected: gradientDirection ==
-                                  selectedGradientDirection,
-                              onGradientDirectionChanged:
-                                  onGradientDirectionChanged),
-                          visible: isCircleRadialButton
-                              ? gradientStyle == GradientStyle.radial
-                              : true,
-                        ),
-                      ),
-                    ],
-                  ),
-                );
-              }).toList()),
-            ],
-          );
-        }).toList());
-  }
-}
-
-class DirectionButton extends StatelessWidget {
-  final IconData icon;
-  final GradientDirection gradientDirection;
-  final bool isSelected;
-  final void Function(GradientDirection) onGradientDirectionChanged;
-
-  const DirectionButton(
-      {Key? key,
-      required this.icon,
-      required this.gradientDirection,
-      required this.isSelected,
-      required this.onGradientDirectionChanged})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final Color greyColor = AppColors.grey;
-    return TextButton(
-      child: Icon(
-        icon,
-        size: 12.0,
-      ),
-      onPressed: () {
-        onGradientDirectionChanged(gradientDirection);
-      },
-      style: TextButton.styleFrom(
-          backgroundColor: isSelected ? greyColor : Colors.white,
-          primary: Colors.black,
-          textStyle: TextStyle(fontWeight: FontWeight.bold),
-          side: BorderSide(
-            color: greyColor,
-          )),
-    );
-  }
-}
-
-class GetGradientButton extends StatefulWidget {
-  const GetGradientButton({
-    Key? key,
-    required this.onTap,
-  }) : super(key: key);
-
-  final Future<void> Function() onTap;
-
-  @override
-  State<GetGradientButton> createState() => _GetGradientButtonState();
-}
-
-class _GetGradientButtonState extends State<GetGradientButton> {
-  bool _showCopiedText = false;
-
-  String get _buttonText => _showCopiedText
-      ? AppStrings.gradientCodeCopied
-      : AppStrings.getGradientCode;
-
-  Color _getBackgroundColor(Set<MaterialState> states) {
-    const Set<MaterialState> interactiveStates = <MaterialState>{
-      MaterialState.pressed,
-      MaterialState.hovered,
-      MaterialState.focused,
-    };
-    if (states.any(interactiveStates.contains)) {
-      return AppColors.darkGrey;
-    }
-    return AppColors.grey;
-  }
-
-  Color _getForegroundColor(Set<MaterialState> states) {
-    const Set<MaterialState> interactiveStates = <MaterialState>{
-      MaterialState.pressed,
-      MaterialState.hovered,
-      MaterialState.focused,
-    };
-    if (states.any(interactiveStates.contains)) {
-      return AppColors.white;
-    }
-    return AppColors.black;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return FractionallySizedBox(
-      widthFactor: 1.0,
-      child: TextButton(
-        child: Text(_buttonText),
-        onPressed: () async {
-          await widget.onTap();
-
-          setState(() {
-            _showCopiedText = true;
-          });
-
-          await Future.delayed(Duration(seconds: 2));
-
-          setState(() {
-            _showCopiedText = false;
-          });
-        },
-        style: ButtonStyle(
-          backgroundColor:
-              MaterialStateProperty.resolveWith(_getBackgroundColor),
-          foregroundColor:
-              MaterialStateProperty.resolveWith(_getForegroundColor),
-          textStyle:
-              MaterialStateProperty.all(TextStyle(fontWeight: FontWeight.bold)),
-          padding: MaterialStateProperty.all(EdgeInsets.all(24)),
-        ),
-      ),
-    );
-  }
-}
